Folder & Routing refactoring
