--- src/components/Contact/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default () => {
-  return (
-    <h1>This will display Contact Us Component</h1>
-  )
-}
