+++ src/components/ContactUs/__tests__/ContactUs.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import ContactUsContainer from '../index'
+import { render } from '@testing-library/react'
+
+describe('Contact Us component', () => {
+  it('should render', () => {
+    const wrapper = render(<ContactUsContainer />)
+    expect(wrapper).toMatchSnapshot();
+  })
+})
