+++ src/components/Header/index.tsx
@@ -5,11 +5,10 @@ import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import IconButton from "@material-ui/core/IconButton";
 import MenuIcon from "@material-ui/icons/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
 import Menu from "@material-ui/core/Menu";
+import NavBar from '../NavBar'
 import { mediaQueries } from '../../mediaqueries'
 import { PetBuddyLogo } from "../Logo";
-import { Link } from 'react-router-dom';
 
 const StyledAppBar = styled(AppBar)`
 display:flex;
@@ -49,48 +48,7 @@ ${mediaQueries("medium")`
 display:none;
 `}
 `
-const StyledMenu = styled(Menu)`
-margin-left: 3.5 %;
-`
-const StyledMenuItem = styled(MenuItem)`
-display:flex;
-flex-direction:column;
-margin:0;
-padding:0 2rem;
-align-items:baseline;
-`
-const NavBarItems = styled.div`
-display: none;
-${mediaQueries("medium")`
-  display:flex;
-  flex-direction:row;
-  justify-content:space-around;
-  width: 60%;
-  margin-left: 20rem;
-  color:#f48fb1;
-  font-weight: 700;
-  `}
-`
-const NavBarLink = styled(Link)`
-text-decoration:none;
-height:2rem;
-opacity: 0.6;
-transition: 0.3s;
-&:hover {
-  margin:0;
-  opacity:1;
-  width:100%;
-  font-weight:700;
-}
-${mediaQueries('laptop')`
-height:auto;
-&:hover {
-  width:auto;
-  background:none;
-  transition: 0.5s;
-  opacity:1;
-}`}
-`
+const StyledMenu = styled(Menu)``
 
 export default function MenuAppBar() {
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
@@ -113,18 +71,18 @@ export default function MenuAppBar() {
         <StyledTypography >
           PetBuddy
           </StyledTypography>
-        <NavBarItems>
-          <NavBarLink to={'/cat-service'}>Cat Services</NavBarLink>
-          <NavBarLink to={'/dog-service'}>Dog Services</NavBarLink>
-          <NavBarLink to={'/pet-service'}>Small Pet Services</NavBarLink>
-          <NavBarLink to={'/contact'}>Contact Us</NavBarLink>
-          <NavBarLink to={'/join'}>Join Us</NavBarLink>
-        </NavBarItems>
+        <NavBar />
         <StyledIconButton
           onClick={handleMenu}>
           <StyledMenuIcon />
         </StyledIconButton>
         <StyledMenu
+          PaperProps={{
+            style: {
+              width: '100vw',
+              textAlign: 'center'
+            },
+          }}
           anchorEl={anchorEl}
           getContentAnchorEl={null}
           anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
@@ -132,13 +90,7 @@ export default function MenuAppBar() {
           open={open}
           onClose={handleClose}
         >
-          <StyledMenuItem className="menu-burguer">
-            <NavBarLink to={'/cat-service'}>Cat Services</NavBarLink>
-            <NavBarLink to={'/dog-service'}>Dog Services</NavBarLink>
-            <NavBarLink to={'/pet-service'}>Small Pet Services</NavBarLink>
-            <NavBarLink to={'/contact'}>Contact Us</NavBarLink>
-            <NavBarLink to={'/join'}>Join Us</NavBarLink>
-          </StyledMenuItem>
+          <NavBar primary />
         </StyledMenu>
       </StyledToolbar>
     </StyledAppBar>
