--- src/components/JoinUs/__tests__/index.test.tsx
@@ -1,1 +1,0 @@
-
