--- src/components/JoinUs/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default () => {
-  return (
-    <h1>Here will go Join our company Component</h1>
-  )
-}
