+++ src/components/NavBar/__tests__/index.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import NavBar from '../index'
+import { render } from '@testing-library/react'
+
+describe('NavBar component', () => {
+  it('should render', () => {
+    const wrapper = render(<NavBar />)
+  })
+  it('should display text', () => {
+    const wrapper = 'Cat Services';
+    expect(wrapper).toEqual('Cat Services')
+  })
+})
