+++ src/screens/Join/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default () => {
+  return (
+    <h1>Here will go JOIN our company Component</h1>
+  )
+}
